tests/wpt/mozilla/tests/mozilla/bluetooth/idl-NavigatorBluetooth.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

test(() => {
    assert_true('bluetooth' in window.navigator);
}, 'navigator.bluetooth IDL test');
</script>
